Restructure codebase into more packages
