--- src/main/java/ExitCommand.java
@@ -1,10 +1,0 @@
-public class ExitCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, Ui ui) {
-        ui.showGoodbye();
-        ui.close();
-    }
-
-    @Override
-    public boolean isExit() { return true; }
-}
