--- src/main/java/MarkCommand.java
@@ -1,13 +1,0 @@
-public class MarkCommand extends Command {
-    private final String index;
-
-    public MarkCommand(String index) {
-        this.index = index;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui) {
-        tasks.mark(index);
-        ui.showTaskMarked(tasks.get(index));
-    }
-}
