+++ src/main/java/vince/command/Command.java
@@ -1,3 +1,9 @@
+package vince.command;
+
+import vince.storage.TaskList;
+import vince.ui.Ui;
+import vince.exception.VinceException;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui) throws VinceException;
     public boolean isExit() { return false; }
