+++ src/main/java/vince/command/OnDateCommand.java
@@ -1,4 +1,9 @@
+package vince.command;
+
 import java.util.List;
+import vince.storage.TaskList;
+import vince.ui.Ui;
+import vince.exception.VinceException;
 
 public class OnDateCommand extends Command {
     private final String dateStr;
@@ -8,7 +13,7 @@ public class OnDateCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, Ui ui) {
+    public void execute(TaskList tasks, Ui ui) throws VinceException {
         String label = tasks.tasksOnDateLabel(dateStr);
         List<String> lines = tasks.tasksOnDateLines(dateStr);
         ui.showTasksOnDate(label, lines);
