+++ src/main/java/vince/exception/VinceException.java
@@ -1,3 +1,5 @@
+package vince.exception;
+
 public class VinceException extends RuntimeException {
     public VinceException(String message) {
         super(message);
