+++ src/main/java/vince/task/Todo.java
@@ -1,3 +1,7 @@
+package vince.task;
+
+import vince.exception.VinceException;
+
 public class Todo extends Task {
     public Todo(String description) throws VinceException {
         super(description);
