+++ text-ui-test/EXPECTED.TXT
@@ -5,11 +5,14 @@ What can I do for you?
 ____________________________________________________________
 
 ____________________________________________________________
+____________________________________________________________
 Got it. I've added this task:
 [T] [ ] read book
 Now you have 1 tasks in the list.
 ____________________________________________________________
 
+____________________________________________________________
+____________________________________________________________
 ____________________________________________________________
 Got it. I've added this task:
 [D] [ ] return book (by: Dec 15 2024, 18:00)
@@ -17,11 +20,15 @@ Now you have 2 tasks in the list.
 ____________________________________________________________
 
 ____________________________________________________________
+____________________________________________________________
+____________________________________________________________
 Got it. I've added this task:
 [E] [ ] project meeting (from: Dec 20 2024, 14:00 to: Dec 20 2024, 16:00)
 Now you have 3 tasks in the list.
 ____________________________________________________________
 
+____________________________________________________________
+____________________________________________________________
 ____________________________________________________________
 Here are the tasks in your list:
 1. [T] [ ] read book
@@ -30,17 +37,24 @@ Here are the tasks in your list:
 ____________________________________________________________
 
 ____________________________________________________________
+____________________________________________________________
+____________________________________________________________
 Tasks on Dec 15 2024:
 2. [D] [ ] return book (by: Dec 15 2024, 18:00)
 ____________________________________________________________
 
+____________________________________________________________
+____________________________________________________________
 ____________________________________________________________
 Tasks on Dec 20 2024:
 3. [E] [ ] project meeting (from: Dec 20 2024, 14:00 to: Dec 20 2024, 16:00)
 ____________________________________________________________
 
 ____________________________________________________________
+____________________________________________________________
+____________________________________________________________
 Bye. Hope to see you again soon!
 
 ____________________________________________________________
 
+____________________________________________________________
